requirements.txt
contourpy==1.3.3
cycler==0.12.1
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.5.1
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.4.1.4
nvidia-cufile-cu12==1.14.1.1
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvshmem-cu12==3.3.9
nvidia-nvtx-cu12==12.9.79
packaging==25.0
pillow==11.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-triton==3.4.0+git11ec6354
setuptools==78.1.0
six==1.17.0
sympy==1.14.0
torch==2.9.0.dev20250729+cu129
torchaudio==2.8.0.dev20250730+cu129
torchvision==0.24.0.dev20250730+cu129
tqdm==4.67.1
typing_extensions==4.14.0